start refactoring to component folders
