+++ public/components/Loader/index.test.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { shallow } from 'enzyme';
-import Loader from './Loader';
+import Loader from './';
 
 describe('Loader', () => {
     it('should render correctly', () => {
